api_secondary_school-master/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
PyJWT==1.7.1
joblib==0.14.0
numpy==1.17.4
matplotlib==3.1.2
mysqlclient
mysql-connector==2.2.9
pandas==1.0.3
python-slugify
scikit-learn==0.23.1
scipy==1.3.3
seaborn==0.9.0
requests==2.22.0
